requirements.txt
tensorflow==2.13.0
numpy==1.23.5
Pillow==10.0.0
requests==2.31.0
opencv-python==4.8.0.76
scikit-learn==1.3.0
matplotlib==3.7.2
